added multiple updates for gamify

+++ frontend/power-planner-fe/src/context/TaskContext.js
@@ -1,21 +1,52 @@
 import axios from "axios";
-import { useState, createContext, useEffect } from "react";
+import { useState, createContext, useEffect, useContext } from "react";
 import * as taskService from "../utils/taskService";
 import { toast } from "react-toastify";
 import useApiCaller from "../hooks/useApiCaller";
+import AuthContext from "../context/AuthContext";
+import { jwtDecode } from "jwt-decode";
+import { callApi } from "../utils/callApi";
 
 const TaskContext = createContext({});
 
 export const TaskProvider = ({ children }) => {
   const [taskList, setTaskList] = useState([]);
+  const { setUserDetails, userDetails } = useContext(AuthContext);
 
+  // Initial user details fetch
+  useEffect(() => {
+    (async () => {
+      const jwt = localStorage.getItem("token");
+      let userId = "";
+
+      if (jwt) {
+        userId = jwtDecode(jwt).id;
+        if (userId) {
+          const verifyUser = await callApi("/home", "GET", {}, jwt);
+          if (verifyUser.status === "success") {
+            setUserDetails(verifyUser.data);
+          }
+        }
+      }
+    })();
+  }, []);
+
+  // Initial Task fetch
   const url = "http://localhost:3003/api/v1/tasks";
   const testUrl = "http://localhost:3000/data";
   const {
-    data: tasks,
+    data: taskData,
     isError: taskError,
     isLoading: tasksLoading,
-  } = useApiCaller(url, "GET", {});
+  } = useApiCaller(url, "GET", {}, userDetails._id);
+
+  // Effects
+  useEffect(() => {
+    if (!tasksLoading && !taskError) {
+      setTaskList(taskData.tasks);
+    }
+    // console.log("Rerender in Context Comp");
+  }, [tasksLoading, taskData]);
 
   // TASK UPDATE HANDL
   const handleTaskUpdate = async (e, taskToUpdate) => {
@@ -29,10 +60,13 @@ export const TaskProvider = ({ children }) => {
     // Update Task on server
     const taskRes = await taskService.updateTask(updatedTask);
 
+    console.log("taskRes", taskRes);
+
     // If task updated
     if (taskRes.status === "success") {
       // update state
       setTaskList(updatedList);
+      setUserDetails(taskRes.data.user);
 
       // If task is completed , show toast
       if (updatedTask.isCompleted)
@@ -50,25 +84,22 @@ export const TaskProvider = ({ children }) => {
   // Task Delete
   const handleTaskDelete = async (taskToDelete) => {
     // console.log("deleted task", taskToDelete);
-    const updatedList = taskList.filter(
-      (task) => task._id !== taskToDelete._id
-    );
-    setTaskList(updatedList);
+
+    if (taskToDelete.isCompleted) {
+      toast.error("Cannot delete a completed task");
+      return;
+    }
     const taskRes = await taskService.deleteTask(taskToDelete);
-    if (taskRes === "success") {
+    if (taskRes.status === "success") {
       toast.success("Task Deleted!");
+      const updatedList = taskList.filter(
+        (task) => task._id !== taskToDelete._id
+      );
+      setTaskList(updatedList);
     } else {
       toast.error("Error! Please try again!");
     }
   };
-
-  // Effects
-  useEffect(() => {
-    if (!tasksLoading && !taskError) {
-      setTaskList(tasks.tasks);
-    }
-    // console.log("Rerender in Context Comp");
-  }, [tasksLoading, tasks]);
 
   return (
     <TaskContext.Provider
